backend/requirements.txt
bcrypt==4.2.1
boto3==1.35.90
easyenvi==1.0.8
fastapi==0.109.2
google-auth-oauthlib==1.2.0
google-auth==2.27.0
google-cloud-aiplatform==1.75.0
html2text==2024.2.26
httpx==0.26.0 
pandas==2.2.3
pillow==11.2.1
pydantic[email]==2.6.1
pyjwt==2.10.1
pymongo==4.10.1
pymupdf==1.25.1
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
regex==2024.11.6
scikit-learn==1.6.0
uvicorn==0.27.1
msal==1.25.0